requirements.txt
Flask>=2.3.2
mysql-connector-python>=8.1.0
Flask-Bcrypt>=1.0.1
requests>=2.31.0
python-telegram-bot==20.4
coinbase-commerce==1.0.1
